appliances: new appliance: vabuilder
This appliance hosts the Gentoo Virtual Appliance[1] tool.  Now you can
build the thing that builds the thing that builds the things.

Note that you cannot (yet) run this inside of a systemd (or Docker)
container (without giving it special privs), but I'm working on that.  It
should at least be possible to build stage4s with it.

[1] https://bitbucket.org/marduk/virtual-appliance